Improve sim-real correspondence
Add Gazebo plugin
This plug in is a modified version of the original ros_control plugin and that plugin's "default hardware plugin" (confusing, I know)
The modified model plugin provides a /runstop service to match the real robot's driver, and hooks this up to ros_controls existing estop implementation
The modified hardware plugin provides the mode switching services that the real robot has (mocked for now) and publishes joint states that include the virtual joints used by the real platform. The wrist extension joint behaves the same as the driver implementation. The rotation and translation joints aren't hooked up yet, and it may be difficult to provide their functionality without forking the actual "follow_joint_trajectory" implementation, which expects only joints that are present in the URDF
Mirror the real platform setup and use a joint state publisher node to republish the subset of joints that match the URDF

Add controller shim that mimics the fake joints used by the real robot

Use effort control for all arm joints
Increase mass of base
Makes tipping behavior in simulation more plausible
I assume that the mass of an assembled base is at least this heavy, given that the robot is around 50lbs total
Use single controller for the simulation
This better matches the physical robot, though this interface doesn't handle the virtual joints that the robot uses

This commit tracks the changes proposed in hello-robot/stretch_ros#51